requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
elevenlabs==0.2.27
python-dotenv==1.0.0 